Renamed to model_input to represent class inside better.
