+++ src/chessbot/model_input.py
